pyproject.toml
[project]
name = "tiku"
version = "0.1.0"
description = "AI 知识库问答代理"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "openai>=1.12.0",
    "python-dotenv>=1.0.0",
    "jieba>=0.42.1",
]

[tool.uv]
